Changed logging to FINEST for AbstractSqlServerRecordSourceApp
FINEST level outputs record ids to standard error and ensures
that Java Hotspot compiler does not optimize away code that
actually reads records. (Without this nominal use of the downloaded
records, the records are otherwise unused, and the Hotspot compiler
detects that the record-reading code can be bypassed without side
effects.

In the various batch and bash scripts, standard error is redirected
to a null oupt stream.